Refactor idle timer to component
